pyproject.toml
[tool.poetry]
name = "vote4film"
version = "1.2.3"
description = "Easy scheduling for regular film nights"
homepage = "https://github.com/Fustra/vote4film/"
repository = "https://github.com/Fustra/vote4film/"
readme = "README.md"
authors = ["QasimK <noreply@QasimK.io>"]
license = "AGPL-3.0"
classifiers = [
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Environment :: No Input/Output (Daemon)",
    "Natural Language :: English",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]
packages = [
    { include = "calender", from = "src"},
    { include = "films", from = "src" },
    { include = "schedule", from = "src" },
    { include = "vote4film", from = "src" },
    { include = "votes", from = "src" },
    { include = "web", from = "src" },
    { include = "manage.py", from = "src" },
]

[tool.poetry.scripts]
manage = 'manage:main'

[tool.poetry.dependencies]
python = "^3.8"
django = "^2.2"
django-environ = "^0.4.5"
django-extensions = "^2.2"
requests = "^2.22"
lxml = "^4.4"
xdg = "^4.0"
psycopg2 = {version = "^2.8", optional = true}
pyhumps = "^1.3.1"
bbfcapi = "^2.0.2"

[tool.poetry.dev-dependencies]
# Checks
black = {version = "=19.10b0", allow-prereleases = true}
isort = "^4.3"
bandit = "^1.6"
pyflakes = "^2.1"
safety = "^1.8"
# Development helpers
django-debug-toolbar = "^2.0"
werkzeug = "^1.0"
# Tests
pytest = "^5.2"
pytest-django = "^3.7"
responses = "^0.10.6"
factory_boy = "^2.12"

[tool.poetry.extras]
postgres = ["psycopg2"]

[tool.black]
target-version = ['py38']

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
